autocraft

+++ src/Timer.cpp
@@ -17,8 +17,8 @@ void Timer::Start()
   mStartTime = std::chrono::high_resolution_clock::now();
 }
 
-double Timer::Elapsed() const
+float Timer::Elapsed() const
 {
   auto endTime = std::chrono::high_resolution_clock::now();
-  return std::chrono::duration<double, std::milli>(endTime - mStartTime).count() / 1000.0;
+  return std::chrono::duration<float, std::milli>(endTime - mStartTime).count() / 1000.0f;
 }
